Planning: openspace: folder refactor (#8294)
