+++ modules/planning/tasks/BUILD
@@ -38,6 +38,11 @@ cc_library(
         "//modules/planning/proto:planning_config_proto",
         "//modules/planning/tasks/deciders:decider",
         "//modules/planning/tasks/deciders/lane_change_decider",
+        "//modules/planning/tasks/deciders/open_space_decider:open_space_fallback_decider",
+        "//modules/planning/tasks/deciders/open_space_decider:open_space_pre_stop_decider",
+        "//modules/planning/tasks/deciders/open_space_decider:open_space_roi_decider",
+        "//modules/planning/tasks/deciders/path_assessment_decider:path_assessment_decider",
+        "//modules/planning/tasks/deciders/path_bounds_decider:path_bounds_decider",
         "//modules/planning/tasks/deciders/path_assessment_decider",
         "//modules/planning/tasks/deciders/path_bounds_decider",
         "//modules/planning/tasks/deciders/path_lane_borrow_decider",
