+++ modules/planning/tasks/deciders/open_space_decider/open_space_pre_stop_decider.cc
@@ -18,7 +18,7 @@
  * @file
  **/
 
-#include "modules/planning/tasks/deciders/open_space_pre_stop_decider.h"
+#include "modules/planning/tasks/deciders/open_space_decider/open_space_pre_stop_decider.h"
 
 #include <algorithm>
 #include <string>
