+++ modules/planning/tasks/deciders/open_space_decider/open_space_pre_stop_decider.h
