+++ modules/planning/tasks/task_factory.cc
@@ -25,9 +25,9 @@
 #include "modules/common/status/status.h"
 #include "modules/planning/tasks/deciders/decider_creep.h"
 #include "modules/planning/tasks/deciders/lane_change_decider/lane_change_decider.h"
-#include "modules/planning/tasks/deciders/open_space_fallback_decider.h"
-#include "modules/planning/tasks/deciders/open_space_pre_stop_decider.h"
-#include "modules/planning/tasks/deciders/open_space_roi_decider.h"
+#include "modules/planning/tasks/deciders/open_space_decider/open_space_fallback_decider.h"
+#include "modules/planning/tasks/deciders/open_space_decider/open_space_pre_stop_decider.h"
+#include "modules/planning/tasks/deciders/open_space_decider/open_space_roi_decider.h"
 #include "modules/planning/tasks/deciders/path_assessment_decider/path_assessment_decider.h"
 #include "modules/planning/tasks/deciders/path_bounds_decider/path_bounds_decider.h"
 #include "modules/planning/tasks/deciders/path_lane_borrow_decider/path_lane_borrow_decider.h"
